agrego verificación al clearTokenRedis

+++ src/utils/clearTokenRedis.js
@@ -3,12 +3,21 @@ require("dotenv").config();
 
 const redisConection = async () => {
   let redisClient;
-  redisClient = createClient({
+
+  const redisOptions = {
     socket: {
       host: process.env.REDIS_HOST,
       port: process.env.REDIS_PORT,
     },
-  });
+  };
+
+  if (process.env.REDIS_PASSWORD !== "null") {
+    redisOptions.password = process.env.REDIS_PASSWORD;
+  }
+
+
+   
+  redisClient = createClient(redisOptions)
 
   redisClient.on("error", (err) => console.log("Redis Client Error", err));
 
@@ -21,7 +30,10 @@ async function deleteToken() {
   const client = await redisConection();
 
   try {
-    // Elimina el token 'at-authjs-token'
+    // Elimina el token 'at-authjs-token' si existe
+    const token = await client.get("at-authjs-token");
+    if (!token) return
+
     await client.del("at-authjs-token");
     console.log('Token "at-authjs-token" destruido en Redis.');
   } catch (err) {
